Another commit for adding things the last one didn't
In this commit the scene which i ran on my phone has
been saved and is being pushed. Removed the rigid
body from the player character which also removed
the problem of incorrect collision due to physics
(as in incorrect for us ) due to the physics
interfering with the head tracking and breaking it
due to incorrect rotation of the camera and it
could potentially disorient users.

The problem with removing the rigid body is movement
now it does not trail off it just stops whenever
we stop using the joystick which leaves an abrupt
feel to it which is not suitable for a character in
flight. Need to find the solution for this.
Also added in this commit are the tree models created
by sanoop and a prefab for the player object.