Adds the snitch logic,new models and final menu level.
This commit adds the AI logic for the snitch.It also
adds the new hogwarts model by sanoop a new compass
model and completed menu level.
The logic for AI is currently chooses a point in
random and moves towards it Still got to work
on the illusion of flight (base offset and move).
music has also been added but a lot of it is still
left plus the UI is also left.