Penultimate push:Almost completes the game.
Adds the final hogwarts asset by sanoop,Beautified the
Terrain of Main level.In it added the fences to hagrids
hut and shifted its place towards the corner .
Merged changes from chinmays scripts into my own scripts.
A level timer has been added that limits the time the player
can stay on the level .Seperated the player controller script
in the menu level from the main level one .It started
giving null pointer errors due to it drawing
references on non-existent objects .
The compass now points towards the snitch after the 10th
ring .
Mainly added the UI for the rings and the snitch to the
player display.Need to scale the UI down so that it appears
correctly on the players screen.